moved SyliusAkeneoLocaleCodeProvider to Provider namespace moved ProductChannelEnabler to Processor namespace
