+++ src/Task/Product/ConfigurableProductsTask.php
@@ -25,13 +25,13 @@ use Synolia\SyliusAkeneoPlugin\Exceptions\Processor\MissingAkeneoAttributeProces
 use Synolia\SyliusAkeneoPlugin\Payload\PipelinePayloadInterface;
 use Synolia\SyliusAkeneoPlugin\Payload\Product\ProductPayload;
 use Synolia\SyliusAkeneoPlugin\Payload\Product\ProductVariantMediaPayload;
+use Synolia\SyliusAkeneoPlugin\Processor\Product\ProductChannelEnablerProcessorInterface;
 use Synolia\SyliusAkeneoPlugin\Provider\AkeneoAttributeProcessorProviderInterface;
 use Synolia\SyliusAkeneoPlugin\Provider\AkeneoTaskProvider;
 use Synolia\SyliusAkeneoPlugin\Repository\ChannelRepository;
 use Synolia\SyliusAkeneoPlugin\Repository\LocaleRepositoryInterface;
 use Synolia\SyliusAkeneoPlugin\Repository\ProductFiltersRulesRepository;
 use Synolia\SyliusAkeneoPlugin\Repository\ProductGroupRepository;
-use Synolia\SyliusAkeneoPlugin\Service\ProductChannelEnabler;
 use Synolia\SyliusAkeneoPlugin\Transformer\ProductOptionValueDataTransformerInterface;
 use Throwable;
 
@@ -85,7 +85,7 @@ final class ConfigurableProductsTask extends AbstractCreateProductEntities
         ProductFiltersRulesRepository $productFiltersRulesRepository,
         AkeneoAttributeProcessorProviderInterface $akeneoAttributeProcessorProvider,
         EventDispatcherInterface $dispatcher,
-        ProductChannelEnabler $productChannelEnabler,
+        ProductChannelEnablerProcessorInterface $productChannelEnabler,
         ProductOptionValueDataTransformerInterface $productOptionValueDataTransformer
     ) {
         parent::__construct(
